Updated Perturbed Plot
Don't worry, I commented out everything previously so it isn't gone. For the perturbed part, we need to plot the departure motion (xpert-x). We were plotting something slightly different I think. This also changes the linearization plot a little that we need to change if we think this is right